Create contiguousSubsetXorProduct.java
Solution To: https://www.hackerrank.com/challenges/sansa-and-xor
Given an array, compute the bitwise or product of contiguous subsets
Runtime: If length of array is even O(1) else O(N). Worst and average case run time O(N).
Explanation: Xor product of a number with itself is 0. Xor product of a number with 0 is itself. 
So all we need to calculate is how many times each number appears in these subsets.
Observation1: Subsets of length X cancels lengths N-X-1. Thus if length is even we get 0.
Observation2: If odd we only need to calculate subsets length ceil(N / 2). We can easily observe
that in that case only even indexed elements are repeated odd times.